+++ src/componentsv2/Proposal/Proposal.jsx
@@ -19,11 +19,7 @@ const Proposal = ({ proposal, extended }) => {
     files,
     name,
     numcomments,
-    publickey,
     publishedat,
-    signature,
-    state,
-    status,
     abandonedat,
     timestamp,
     userid,
