+++ src/containers/User/Detail/Proposals.jsx
@@ -1,11 +1,7 @@
 import React from "react";
 
 const Proposals = () => {
-  return (
-    <div>
-      Proposals
-    </div>
-  );
+  return <div>Proposals</div>;
 };
 
 export default Proposals;
